move pack and unpack to new encoding package
